+++ code/system.ui/react.dev/src/types.ts
@@ -1,3 +1,4 @@
 export type * from './types.def';
-export type * from './ui/List.Spec/t';
+export type * from './ui/Harness/t';
 export type * from './ui/List.Module/t';
+export type * from './ui/List.Spec/t';
